server: moved mongo id in FlightPlan to a new version of struct named DBFlightPlan I ran into a problem with FlightPlan having an ID field. I thought it would be okay to just leave this field blank and have it be omitted when empty. The problem was that the ID field was never actually empty, but instead was just initialized to all zeros. So the ID field was actually never omitted. This led to a weird bug where inserting FlightPlan's into mongodb created a duplicate key error, bc ID was all zeros in 2 inserts, so it created a document with ID 0, and then tried to insert a new one and saw a document with this ID already existed.
The fix was to move the ID field to a new struct DBFlightPlan which has FlightPlan as an embedded struct. This way when you want to have the mongodb ID present you have to explicitly use DBFlightPlan. I think this is a good change because before it was ambiguous when ID was and wasn't going to be set to a meaningful value.